requirements.txt
discord.py
python-dotenv
supabase
numpy
yfinance
emoji
pyyaml
